Refactor training pipeline and finish SwinV2 removal cleanup
- Strip remaining SwinV2 config/schema/training_config remnants now that
  ViT (SimplifiedTagger) is the only supported architecture
- Add explicit data.color_order (RGB/BGR) handling with channel-order-aware
  normalization guidance
- Trim training_config.py (~830 lines) and schemas.py to match the ViT-only
  surface; refactor dataset_loader, train_direct, training_utils,
  validation_loop, and evaluation_metrics
- Misc cleanup across config, ONNX inference, metadata, and utils

+++ custom_drop_path.py
@@ -15,6 +15,8 @@ class SafeDropPath(nn.Module):
         keep_prob = 1.0 - self.drop_prob
         # keep_prob is always > 0 due to __init__ validation
         shape = (x.shape[0],) + (1,) * (x.ndim - 1)
-        random_tensor = keep_prob + torch.rand(shape, dtype=x.dtype, device=x.device)
+        # Draw in float32: drawing in x.dtype (e.g. bf16) quantizes keep_prob
+        # and skews the realized keep rate. The 0/1 mask casts back exactly.
+        random_tensor = keep_prob + torch.rand(shape, dtype=torch.float32, device=x.device)
         random_tensor.floor_()
-        return x.div(keep_prob) * random_tensor
+        return x.div(keep_prob) * random_tensor.to(x.dtype)
